# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  controller.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.AreaCarrera:
    properties:
      carrerasTotales:
        items:
          $ref: '#/definitions/models.CarreraTotal'
        type: array
      cod_carea_carrera:
        type: integer
      nom_area_carrera:
        type: string
    type: object
  models.CargaParcial:
    properties:
      ano:
        type: integer
      cant_grupos_practica:
        type: integer
      cant_grupos_teoria:
        type: integer
      cargasTotales:
        items:
          $ref: '#/definitions/models.CargaTotal'
        type: array
      cod_curso_dictar:
        type: integer
      hrs_crono_presen_semestral:
        type: number
      hrs_ejercicio:
        type: number
      hrs_practica:
        type: number
      hrs_teoria:
        type: number
      id_carga_parcial:
        type: integer
      planes:
        items:
          $ref: '#/definitions/models.Plan'
        type: array
      sem:
        type: integer
      tipo_curso:
        type: integer
      total_hrs_docencia:
        type: number
    type: object
  models.CargaTotal:
    properties:
      cantidad_estimada_alum:
        type: integer
      cargaParcial:
        $ref: '#/definitions/models.CargaParcial'
      cargasTotalesProfesores:
        items:
          $ref: '#/definitions/models.CargaTotalProfesor'
        type: array
      cod_tipo_seccion:
        type: integer
      hcp:
        type: integer
      id_carga_parcial:
        type: integer
      id_carga_total:
        type: integer
      nom_grupo:
        type: string
      num_grupos:
        type: integer
      tipoSeccion:
        $ref: '#/definitions/models.TipoSeccion'
      usuariosAulas:
        items:
          $ref: '#/definitions/models.UsuarioAula'
        type: array
    type: object
  models.CargaTotalProfesor:
    properties:
      cargaTotal:
        $ref: '#/definitions/models.CargaTotal'
      cod_estado_carga:
        type: integer
      cod_persona:
        type: string
      id_carga_total:
        type: integer
      id_carga_total_profesor:
        type: integer
      tipo_carga:
        type: integer
    type: object
  models.Carrera:
    properties:
      carrerasTotales:
        items:
          $ref: '#/definitions/models.CarreraTotal'
        type: array
      cod_carrera:
        type: string
      nom_carrera:
        type: string
    type: object
  models.CarreraTotal:
    properties:
      ano_cierre:
        type: integer
      ano_creacion:
        type: integer
      area_conocimiento_cna_id:
        type: integer
      areaCarrera:
        $ref: '#/definitions/models.AreaCarrera'
      carrera:
        $ref: '#/definitions/models.Carrera'
      centro:
        $ref: '#/definitions/models.CentroUniversitario'
      cod_area_carrera:
        type: integer
      cod_carrera:
        type: string
      cod_centro:
        type: integer
      cod_curso_dictar:
        type: integer
      cod_especialidad:
        type: integer
      cod_mencion:
        type: integer
      cod_propietario:
        type: integer
      codigo_sies:
        type: string
      especialidad:
        $ref: '#/definitions/models.Especialidad'
      facDep:
        $ref: '#/definitions/models.FacDep'
      id_carrera_total:
        type: integer
      id_fac_dep:
        type: integer
      mencion:
        $ref: '#/definitions/models.Mencion'
      nivelCarrera:
        $ref: '#/definitions/models.NivelCarrera'
      nro_decreto_creacion:
        type: string
      planes:
        items:
          $ref: '#/definitions/models.Plan'
        type: array
      propietarioCarrera:
        $ref: '#/definitions/models.PropietarioCarrera'
      regimen:
        type: string
      vigencia:
        type: integer
    type: object
  models.CentroUniversitario:
    properties:
      ano_cierre:
        type: integer
      ano_inicio:
        type: integer
      carrerasTotales:
        items:
          $ref: '#/definitions/models.CarreraTotal'
        type: array
      cod_centro:
        type: integer
      cod_resolucion_creacion:
        type: string
      nom_centro:
        type: string
    type: object
  models.CompetenciaAsignatura:
    properties:
      id:
        type: integer
      nom_comp_asig:
        type: string
    type: object
  models.CompetenciaRc:
    properties:
      id:
        type: integer
      nom_competencia:
        type: string
    type: object
  models.CursoComun:
    properties:
      cargaParcial:
        $ref: '#/definitions/models.CargaParcial'
      cod_tipo_curso_dictar:
        type: integer
      cod_unidad_responsable:
        type: integer
      detallesCursosComunes:
        items:
          $ref: '#/definitions/models.DetalleCursoComun'
        type: array
      hrs_crono_presen:
        type: number
      id_curso_comun:
        type: integer
      nom_curso_comun:
        type: string
    type: object
  models.CursoDictar:
    properties:
      ano:
        type: integer
      cargasParciales:
        items:
          $ref: '#/definitions/models.CargaParcial'
        type: array
      cod_estado_curso:
        type: integer
      cod_unidad_responsable:
        type: integer
      detallesCursosComunes:
        items:
          $ref: '#/definitions/models.DetalleCursoComun'
        type: array
      hrs_crono_presen:
        type: number
      id:
        type: integer
      id_malla:
        type: integer
      modulo:
        type: integer
      se_dicta:
        type: integer
      sem:
        type: integer
    type: object
  models.CursoPionero:
    properties:
      cod_tipo_modulo:
        type: integer
      competenciaAsignatura:
        $ref: '#/definitions/models.CompetenciaAsignatura'
      cursosPionerosCompetencias:
        items:
          $ref: '#/definitions/models.CursoPioneroCompetencia'
        type: array
      es_TEL:
        type: integer
      hrs_crono_no_presen_semestral:
        type: number
      hrs_crono_presen_semestral:
        type: number
      hrs_ejercicios:
        type: number
      hrs_practicas:
        type: number
      hrs_teoricas:
        type: number
      id_comp_asig:
        type: integer
      id_curso_pionero:
        type: integer
      id_nivel_comp:
        type: integer
      institucional:
        type: integer
      modsubsAsigcoms:
        items:
          $ref: '#/definitions/models.ModsubAsigcom'
        type: array
      modulosSubcompetencias:
        items:
          $ref: '#/definitions/models.ModuloSubcompetencia'
        type: array
      sct_total:
        type: number
      telsScts:
        items:
          $ref: '#/definitions/models.TelSct'
        type: array
    type: object
  models.CursoPioneroCompetencia:
    properties:
      competenciaRc:
        $ref: '#/definitions/models.CompetenciaRc'
      id_competencia:
        type: integer
      id_curso_pionero:
        type: integer
      id_curso_pionero_comp:
        type: integer
      id_nivel_comp:
        type: integer
    type: object
  models.DetalleCursoComun:
    properties:
      cod_tipo_plan:
        type: integer
      cursoComun:
        $ref: '#/definitions/models.CursoComun'
      cursoDictar:
        $ref: '#/definitions/models.CursoDictar'
      idCursoComun:
        type: integer
      idCursoDictar:
        type: integer
      tipoPlan:
        $ref: '#/definitions/models.TipoPlan'
    type: object
  models.Especialidad:
    properties:
      carrerasTotales:
        items:
          $ref: '#/definitions/models.CarreraTotal'
        type: array
      id:
        type: integer
      nom_especialidad:
        type: string
    type: object
  models.EstadoFinalMatricula:
    properties:
      descripcion:
        type: string
      id:
        type: integer
      matriculas:
        items:
          $ref: '#/definitions/models.Matricula'
        type: array
    type: object
  models.EstadoInscripcion:
    properties:
      id:
        type: integer
      nom_estado_inscripcion:
        type: string
    type: object
  models.FacDep:
    properties:
      carrerasTotales:
        items:
          $ref: '#/definitions/models.CarreraTotal'
        type: array
      cod_unidad_depto:
        type: integer
      cod_unidad_facultad:
        type: integer
      id:
        type: integer
    type: object
  models.MallaPionero:
    properties:
      cod_tipo_curso_malla:
        type: integer
      curso_basico:
        type: integer
      cursoPionero:
        $ref: '#/definitions/models.CursoPionero'
      cursosDictar:
        items:
          $ref: '#/definitions/models.CursoDictar'
        type: array
      duracion:
        type: string
      es_proyecto:
        type: integer
      frecuencia_a_dictar:
        type: integer
      id:
        type: integer
      id_curso_pionero:
        type: integer
      id_modsub:
        type: integer
      id_planes:
        type: integer
      incluye_grado:
        type: integer
      mallasRequisitos:
        items:
          $ref: '#/definitions/models.MallaRequisito'
        type: array
      nro_ano:
        type: integer
      nro_sem:
        type: integer
      plan:
        $ref: '#/definitions/models.Plan'
      tipoCursoMalla:
        $ref: '#/definitions/models.TipoCursoMalla'
    type: object
  models.MallaRequisito:
    properties:
      id:
        type: integer
      id_malla:
        type: integer
      id_malla_req:
        type: integer
      mallaPionero:
        $ref: '#/definitions/models.MallaPionero'
      tipo:
        type: integer
    type: object
  models.Matricula:
    properties:
      ano_ingreso:
        type: integer
      ano_matricula:
        type: integer
      ano_primer_ingreso:
        type: integer
      ano_proceso:
        type: integer
      cod_estado_final:
        type: integer
      cod_modo_ingreso:
        type: integer
      cod_persona:
        type: integer
      cod_tipo_alumno:
        type: integer
      cod_tipo_matricula:
        type: integer
      estadoFinalMatricula:
        $ref: '#/definitions/models.EstadoFinalMatricula'
      id_matricula:
        type: integer
      id_planes:
        type: integer
      plan:
        $ref: '#/definitions/models.Plan'
      rut_alum:
        type: string
      sem_desvinculacion:
        type: integer
    type: object
  models.Mencion:
    properties:
      carrerasTotales:
        items:
          $ref: '#/definitions/models.CarreraTotal'
        type: array
      id:
        type: integer
      nom_mencion:
        type: string
    type: object
  models.ModsubAsigcom:
    properties:
      cursoPionero:
        $ref: '#/definitions/models.CursoPionero'
      idCursoPionero:
        type: integer
      idModsub:
        type: integer
    type: object
  models.ModuloSubcompetencia:
    properties:
      cod_tipo_modulo:
        type: integer
      competenciasRc:
        items:
          $ref: '#/definitions/models.CompetenciaRc'
        type: array
      hrs_a_guiado:
        type: number
      hrs_no_presenciales:
        type: number
      hrs_presenciales:
        type: number
      id:
        type: integer
      institucional:
        type: integer
      nom_modsub:
        type: string
      sct:
        type: integer
    type: object
  models.NivelCarrera:
    properties:
      carrerasTotales:
        items:
          $ref: '#/definitions/models.CarreraTotal'
        type: array
      descripcion:
        type: string
      id:
        type: integer
      planes:
        items:
          $ref: '#/definitions/models.Plan'
        type: array
    type: object
  models.NivelGlobalCarrera:
    properties:
      descripcion:
        type: string
      id:
        type: integer
      planes:
        items:
          $ref: '#/definitions/models.Plan'
        type: array
    type: object
  models.Plan:
    properties:
      anoPlan:
        type: integer
      caracteristicas:
        type: string
      cargaParcial:
        $ref: '#/definitions/models.CargaParcial'
      carreraTotal:
        $ref: '#/definitions/models.CarreraTotal'
      cod_jornada:
        type: integer
      cod_nivel_carrera:
        type: integer
      cod_nivel_global:
        type: integer
      cod_regimen:
        type: integer
      cod_tipo_plan:
        type: integer
      codModalidad:
        type: integer
      duracion:
        type: integer
      id:
        type: integer
      id_carga_parcial:
        type: integer
      id_carrera_total:
        type: integer
      mallasPioneros:
        items:
          $ref: '#/definitions/models.MallaPionero'
        type: array
      matriculas:
        items:
          $ref: '#/definitions/models.Matricula'
        type: array
      nivelCarrera:
        $ref: '#/definitions/models.NivelCarrera'
      nivelGlobalCarrera:
        $ref: '#/definitions/models.NivelGlobalCarrera'
      regimen:
        $ref: '#/definitions/models.Regimen'
      tipoPlan:
        $ref: '#/definitions/models.TipoPlan'
      vigencia:
        type: integer
    type: object
  models.PropietarioCarrera:
    properties:
      carrerasTotales:
        items:
          $ref: '#/definitions/models.CarreraTotal'
        type: array
      id:
        type: integer
      nom_propietario:
        type: string
    type: object
  models.Regimen:
    properties:
      id:
        type: integer
      nom_regimen:
        type: string
      vigente:
        type: integer
    type: object
  models.TelSct:
    properties:
      sct_P:
        type: integer
      tel_P:
        type: integer
    type: object
  models.TipoCursoMalla:
    properties:
      descripcion:
        type: string
      id:
        type: integer
    type: object
  models.TipoPlan:
    properties:
      detallesCursosComunes:
        items:
          $ref: '#/definitions/models.DetalleCursoComun'
        type: array
      id:
        type: integer
      nom_tipo_plan:
        type: string
      planes:
        items:
          $ref: '#/definitions/models.Plan'
        type: array
    type: object
  models.TipoSeccion:
    properties:
      id:
        type: integer
      nom_tipo_seccion:
        type: string
    type: object
  models.TipoUsuario:
    properties:
      id:
        type: integer
      nom_tipo_usuario:
        type: string
    type: object
  models.UsuarioAula:
    properties:
      ano:
        type: integer
      cod_estado_inscripcion:
        type: integer
      cod_persona:
        type: integer
      estadoInscripcion:
        $ref: '#/definitions/models.EstadoInscripcion'
      id_carga_total:
        type: integer
      id_malla:
        type: integer
      id_malla_cd:
        type: integer
      id_tipo_usuario:
        type: integer
      id_usuario_aula:
        type: integer
      mallaPionero:
        $ref: '#/definitions/models.MallaPionero'
      opcion:
        type: integer
      sem:
        type: integer
      tipoUsuario:
        $ref: '#/definitions/models.TipoUsuario'
    type: object
host: backend.localhost
info:
  contact:
    email: citiaps@usach.cl
    name: CITIAPS
    url: https://citiaps.cl
  description: Este servicio entrega las mallas y planes de estudio de un estudiante
    de la Universidad de Magallanes.
  license:
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Servicio de Mallas
  version: "1.0"
paths:
  /mallas/cursos_complementarios/{id}:
    get:
      consumes:
      - application/json
      description: Esta función obtiene la malla con cursos complementarios del usuario
      parameters:
      - description: id de la malla a consultar
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Consulta realizada con éxito
          schema:
            items:
              $ref: '#/definitions/models.MallaPionero'
            type: array
        "404":
          description: No se encontró datos.
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtiene la malla con cursos complementarios del usuario autenticado
      tags:
      - Malla
  /mallas/pdf/malla/{id}:
    get:
      consumes:
      - application/json
      description: Esta función genera un archivo PDF con la malla del usuario autenticado
        en base al Bearer Token que viene en el encabezado del request.
      parameters:
      - description: id de la malla a consultar
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Un archivo PDF con la malla del usuario autenticado.
          schema:
            type: file
        "404":
          description: No se encontró datos.
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - Bearer: []
      summary: Genera un archivo PDF con la malla del usuario autenticado
      tags:
      - Pdf
  /mallas/plan/:
    get:
      consumes:
      - application/json
      description: Esta función obtiene todos los planes de estudio.
      responses:
        "200":
          description: Consulta realizada con éxito
          schema:
            items:
              $ref: '#/definitions/models.Plan'
            type: array
        "404":
          description: No se encontró datos.
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtiene todos los planes de estudio
      tags:
      - Plan
  /mallas/plan/carreras/:
    get:
      consumes:
      - application/json
      description: Esta función obtiene los planes de estudio del usuario autenticado
        en base al Bearer Token que viene en el encabezado del request.
      responses:
        "200":
          description: Consulta realizada con éxito
          schema:
            items:
              $ref: '#/definitions/models.Plan'
            type: array
        "404":
          description: No se encontró datos.
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtiene los planes de estudio del usuario autenticado
      tags:
      - Plan
  /mallas/plan/plan/{id}:
    get:
      consumes:
      - application/json
      description: Esta función obtiene un plan de estudio en base a su id.
      parameters:
      - description: id del plan a consultar
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Consulta realizada con éxito
          schema:
            items:
              $ref: '#/definitions/models.Plan'
            type: array
        "404":
          description: No se encontró datos.
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtiene un plan de estudio
      tags:
      - Plan
  /mallas/sin_cursos_complementarios/{id}:
    get:
      consumes:
      - application/json
      description: Esta función obtiene la malla sin cursos complementarios del usuario
        autenticado en base al Bearer Token que viene en el encabezado del request.
      parameters:
      - description: id de la malla a consultar
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Consulta realizada con éxito
          schema:
            items:
              $ref: '#/definitions/models.MallaPionero'
            type: array
        "404":
          description: No se encontró datos.
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      security:
      - Bearer: []
      summary: Obtiene la malla sin cursos complementarios del usuario autenticado
      tags:
      - Malla
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: BearerJWTToken in Authorization Header
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"